Keep Veryfront adapter tests easier to extend
The Veryfront adapter test suite repeated adapter construction, cache seeding, and async wait helpers inline, so this extracts the shared test scaffolding into a sibling helper module without changing adapter assertions.

Constraint: Refactor must preserve adapter test behavior and cache warmup assertions exactly
Constraint: veryfront-code pre-commit verify:quick currently fails on pre-existing CLI boundary violations in cli/commands/extension/*.ts
Rejected: Extract client stubbing helpers too | would broaden this pass beyond the narrow shared test scaffolding targeted by the wave
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep reusable Veryfront adapter test scaffolding in the helper module unless a future case needs bespoke inline setup
Tested: deno test --no-check --allow-all src/platform/adapters/fs/veryfront/adapter.test.ts; deno fmt --check src/platform/adapters/fs/veryfront/adapter.test.ts src/platform/adapters/fs/veryfront/adapter.test-helpers.ts; deno lint src/platform/adapters/fs/veryfront/adapter.test.ts src/platform/adapters/fs/veryfront/adapter.test-helpers.ts; git diff --check
Not-tested: Full veryfront-code repo suite
